jawthorizer/pyproject.toml
[tool.poetry]
name = "jawthorizer"
version = "1.0.0+dev"
description = "Provides an authorizer for AWS lambda"
authors = ["OmniVector SOlutions <info@omnivector.solutions>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.18.18"
PyJWT = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
pytest-env = "^0.6.2"
python-lambda-local = "^0.1.12"
pytest-random-order = "^1.0.4"
black = "^21.7b0"
isort = "^5.9.3"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--random-order --cov=jawthorizer --cov-report=term-missing --cov-fail-under=94"
testpaths = ["tests"]
env = [
    "JAWTHORIZER_APP_SHORT_NAME = my-app",
    "JAWTHORIZER_STAGE = edge",
    "JAWTHORIZER_REGION = us-west-2",
    "JAWTHORIZER_ALLOWED_SUBS = '*'",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
